+++ .github/workflows/ci.yml
@@ -25,6 +25,6 @@ jobs:
           pip install .
 
       - name: tests
-        run: python -m pytest --cov vipy --cov-report=xml
+        run: python -m pytest --cov pyvisgen --cov-report=xml
 
       - uses: codecov/codecov-action@v1
