+++ pyvisgen/layouts/vlba_light.txt
