+++ tests/test_layouts.py
@@ -2,7 +2,7 @@ import numpy as np
 
 
 def test_get_array_layout():
-    from vipy.layouts.layouts import get_array_layout
+    from pyvisgen.layouts.layouts import get_array_layout
 
     layout = get_array_layout("eht")
